// src/services/publish/typings.d.ts
/**
 * Copyright 2024 Maner·Fan
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import '@interface/typings';

declare namespace PUBLISH {
  type PublishConfigEntity = {
    /**
     * uid
     */
    uid?: string;

    /**
     * 创建时间
     */
    created_at?: number;

    /**
     * 更新时间
     */
    updated_at?: number;
    /**
     * 目标类型
     */
    target_type?: string;

    /**
     * 目标 ID
     */
    target_uid?: string;

    /**
     * 配置
     */
    config?: Record<string, any>;

    /**
     * 发布状态
     */
    publish_status?: "DRAFT" | "PUBLISHED";
  };
}